+++ ui/app/src/components/AuditGridFilterPopover/AuditGridFilterPopover.tsx
@@ -14,29 +14,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { ChangeEvent, useCallback, useMemo, useState } from 'react';
+import React from 'react';
 import Popover from '@material-ui/core/Popover';
-import styles from './styles';
-import TextField from '@material-ui/core/TextField';
-import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
-import MenuItem from '@material-ui/core/MenuItem';
+import useStyles from './styles';
 import { Site } from '../../models/Site';
 import { PagedArray } from '../../models/PagedArray';
 import User from '../../models/User';
-import { useDebouncedInput } from '../../utils/hooks';
-import Box from '@material-ui/core/Box';
-import moment from 'moment-timezone';
-import { MaterialUiPickersDate } from '@material-ui/pickers/typings/date';
-import Autocomplete from '@material-ui/lab/Autocomplete';
-import { KeyboardDatePicker, MuiPickersUtilsProvider } from '@material-ui/pickers';
-import DateFnsUtils from '@date-io/date-fns';
-import Tooltip from '@material-ui/core/Tooltip';
-import IconButton from '@material-ui/core/IconButton';
-import ClearRoundedIcon from '@material-ui/icons/ClearRounded';
-import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
+import AuditGridFilterPopoverBody from './AuditGridFilterPopoverBody';
 
-interface AuditGridFilterPopoverProps {
+export interface AuditGridFilterPopoverProps {
   open: boolean;
   filterId: 'operationTimestamp' | 'siteId' | 'user' | 'origin' | 'operations' | 'target' | 'clusterNodeId';
   anchorPosition: {
@@ -60,40 +46,9 @@ interface AuditGridFilterPopoverProps {
   onFilterChange(fieldId: string, value: any): void;
 }
 
-const translations = defineMessages({
-  siteId: {
-    id: 'auditGridFilterPopover.filterBySite',
-    defaultMessage: 'Filter by Site'
-  },
-  user: {
-    id: 'auditGridFilterPopover.filterByUser',
-    defaultMessage: 'Filter by User'
-  },
-  origin: {
-    id: 'auditGridFilterPopover.filterByOrigin',
-    defaultMessage: 'Filter by Origin'
-  },
-  operations: {
-    id: 'auditGridFilterPopover.filterByOperations',
-    defaultMessage: 'Filter by Operations'
-  },
-  target: {
-    id: 'auditGridFilterPopover.filterByTarget',
-    defaultMessage: 'Filter by Target Value'
-  },
-  clusterNodeId: {
-    id: 'auditGridFilterPopover.filterByCluster',
-    defaultMessage: 'Filter by Cluster Node'
-  },
-  allOperations: {
-    id: 'auditGridFilterPopover.allOperations',
-    defaultMessage: 'All Operations'
-  }
-});
-
 export default function AuditGridFilterPopover(props: AuditGridFilterPopoverProps) {
   const { open, anchorPosition, onClose } = props;
-  const classes = styles();
+  const classes = useStyles();
   return (
     <Popover
       open={open}
@@ -106,236 +61,7 @@ export default function AuditGridFilterPopover(props: AuditGridFilterPopoverProp
         horizontal: 'center'
       }}
     >
-      <AuditGridFilterPopoverContainer {...props} />
+      <AuditGridFilterPopoverBody {...props} />
     </Popover>
-  );
-}
-
-export function AuditGridFilterPopoverContainer(props: AuditGridFilterPopoverProps) {
-  const { filterId, value, onFilterChange, timezone, onTimezoneSelected, onResetFilter, onClose } = props;
-  const classes = styles();
-  const { sites, users, operations, origins, timezones } = props.options;
-  const { formatMessage } = useIntl();
-  const [keyword, setKeyword] = useState(props.value ?? '');
-  const [fromDate, setFromDate] = useState(props.dateFrom ? moment(props.dateFrom) : null);
-  const [toDate, setToDate] = useState(props.dateTo ? moment(props.dateTo) : null);
-  const onSearch = useCallback((keywords: string) => onFilterChange(filterId, keywords), [onFilterChange, filterId]);
-
-  const onSearch$ = useDebouncedInput(onSearch, 400);
-
-  const options = useMemo<{ id: string; value: string; name: string | React.ReactNode }[]>(() => {
-    switch (filterId) {
-      case 'siteId': {
-        return [
-          {
-            id: 'all',
-            value: 'all',
-            name: <FormattedMessage id="auditGrid.allSites" defaultMessage="All Sites" />
-          },
-          {
-            id: 'studio_root',
-            value: 'studio_root',
-            name: <FormattedMessage id="words.system" defaultMessage="System" />
-          },
-          ...sites.map((site) => ({ id: site.id, name: site.name, value: site.id }))
-        ];
-      }
-      case 'user': {
-        return [
-          {
-            id: 'all',
-            value: 'all',
-            name: <FormattedMessage id="auditGrid.allUsers" defaultMessage="All Users" />
-          },
-          ...users.map((user) => ({ id: user.id.toString(), name: user.username, value: user.username }))
-        ];
-      }
-      case 'origin': {
-        return [
-          {
-            id: 'all',
-            value: 'all',
-            name: <FormattedMessage id="auditGrid.allOrigins" defaultMessage="All Origins" />
-          },
-          ...origins
-        ];
-      }
-      case 'operations': {
-        return [
-          {
-            id: 'all',
-            value: 'all',
-            name: formatMessage(translations.allOperations)
-          },
-          ...operations
-        ];
-      }
-    }
-  }, [filterId, formatMessage, operations, origins, sites, users]);
-
-  const onTextFieldChanges = (e: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => {
-    onSearch$.next(e.target.value);
-    setKeyword(e.target.value);
-  };
-
-  const onMultipleSelectChanges = (e: React.ChangeEvent<{ value: any }>) => {
-    const lastString = e.target.value[e.target.value.length - 1];
-    if (lastString === 'all' || e.target.value.length === 0) {
-      onFilterChange(filterId, 'all');
-    } else {
-      const values = e.target.value.filter((value) => value !== 'all');
-      onFilterChange(filterId, values.join());
-    }
-  };
-
-  const onFromDateSelected = (date: MaterialUiPickersDate) => {
-    onFilterChange('dateFrom', date ? moment(date).format() : 'all');
-    setFromDate(date);
-  };
-
-  const onToDateSelected = (date: MaterialUiPickersDate) => {
-    onFilterChange('dateTo', date ? moment(date).format() : 'all');
-    setToDate(date);
-  };
-
-  const onClearDates = () => {
-    setToDate(null);
-    setFromDate(null);
-    onResetFilter(['dateFrom', 'dateTo']);
-  };
-
-  const onClearTextField = () => {
-    onResetFilter(filterId);
-    setKeyword('');
-  };
-
-  return (
-    <>
-      <Avatar className={classes.popoverCloseIcon} onClick={onClose}>
-        <ClearRoundedIcon fontSize="small" />
-      </Avatar>
-      {filterId === 'operationTimestamp' && (
-        <MuiPickersUtilsProvider utils={DateFnsUtils}>
-          <form className={classes.popoverForm} noValidate autoComplete="off">
-            <Box display="flex" alignItems="center" marginBottom="20px">
-              <KeyboardDatePicker
-                className={classes.fromDatePicker}
-                clearable
-                margin="none"
-                label={<FormattedMessage id="words.from" defaultMessage="From" />}
-                format="MM/dd/yyyy"
-                value={fromDate}
-                onChange={onFromDateSelected}
-              />
-              <KeyboardDatePicker
-                className={classes.toDatePicker}
-                clearable
-                margin="none"
-                label={<FormattedMessage id="words.to" defaultMessage="To" />}
-                format="MM/dd/yyyy"
-                value={toDate}
-                onChange={onToDateSelected}
-              />
-              <Button
-                className={classes.clearButton}
-                disabled={!toDate && !fromDate}
-                variant="text"
-                color="primary"
-                onClick={() => onClearDates()}
-              >
-                <FormattedMessage id="words.clear" defaultMessage="Clear" />
-              </Button>
-            </Box>
-            <Autocomplete
-              disableClearable
-              options={timezones}
-              getOptionLabel={(option) => option}
-              value={timezone}
-              onChange={(e: React.ChangeEvent<{}>, value) => {
-                onTimezoneSelected(value);
-              }}
-              fullWidth
-              renderInput={(params) => (
-                <TextField
-                  {...params}
-                  label={<FormattedMessage id="auditGrid.timezone" defaultMessage="Timezone" />}
-                  variant="outlined"
-                />
-              )}
-            />
-          </form>
-        </MuiPickersUtilsProvider>
-      )}
-      {['siteId', 'user', 'origin'].includes(filterId) && (
-        <Box display="flex" alignItems="center">
-          <TextField
-            fullWidth
-            select
-            label={formatMessage(translations[filterId])}
-            value={value ? value : 'all'}
-            onChange={(e) => onFilterChange(filterId, e.target.value)}
-          >
-            {options.map((option) => (
-              <MenuItem key={option.id} value={option.value}>
-                {option.name}
-              </MenuItem>
-            ))}
-          </TextField>
-          <Button
-            className={classes.clearButton}
-            disabled={!value}
-            variant="text"
-            color="primary"
-            onClick={() => onResetFilter(filterId)}
-          >
-            <FormattedMessage id="words.clear" defaultMessage="Clear" />
-          </Button>
-        </Box>
-      )}
-      {filterId === 'operations' && (
-        <Box display="flex" alignItems="center">
-          <TextField
-            fullWidth
-            select
-            label={formatMessage(translations[filterId])}
-            value={value?.split(',') ?? ['all']}
-            SelectProps={{ multiple: true }}
-            onChange={onMultipleSelectChanges}
-          >
-            {options.map((option) => (
-              <MenuItem key={option.id} value={option.value}>
-                {option.name}
-              </MenuItem>
-            ))}
-          </TextField>
-          <Button
-            className={classes.clearButton}
-            disabled={!value}
-            variant="text"
-            color="primary"
-            onClick={() => onResetFilter(filterId)}
-          >
-            <FormattedMessage id="words.clear" defaultMessage="Clear" />
-          </Button>
-        </Box>
-      )}
-      {['target', 'clusterNodeId'].includes(filterId) && (
-        <TextField
-          value={keyword}
-          label={formatMessage(translations[filterId])}
-          InputProps={{
-            endAdornment: keyword && (
-              <Tooltip title={<FormattedMessage id="words.clear" defaultMessage="Clear" />}>
-                <IconButton size="small" className={classes.clearButton} onClick={() => onClearTextField()}>
-                  <ClearRoundedIcon />
-                </IconButton>
-              </Tooltip>
-            )
-          }}
-          fullWidth
-          onChange={onTextFieldChanges}
-        />
-      )}
-    </>
   );
 }
