+++ ui/app/src/models/Audit.ts
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export interface LogParameters {
+export interface AuditLogEntryParameter {
   targetId: string;
   targetType: string;
   targetSubtype: string;
@@ -36,5 +36,5 @@ export interface AuditLogEntry {
   actorId: string;
   actorDetails: string;
   clusterNodeId: string;
-  parameters: LogParameters[];
+  parameters: AuditLogEntryParameter[];
 }
